perf: opt inspector
