+++ web-packages/imgui-demo/src/widgets/widget.ts
@@ -1,0 +1,5 @@
+export class Widget {
+  draw () {
+
+  }
+}
